backend/src/interview-ai/requirements.txt
# Flask & Web
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Computer Vision
opencv-python>=4.9.0
ultralytics>=8.0.0
mediapipe>=0.10.0

# Machine Learning - Versions compatibles CRITIQUES
tensorflow>=2.13.0,<2.17.0
protobuf>=3.20.3,<4.21.0          # ← FIX: Version compatible avec TensorFlow
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.35.0

# Audio
librosa>=0.10.0
soundfile>=0.12.0

# Data & Serialization
numpy>=1.24.0,<2.0.0               # ← Important: <2.0 pour compatibilité pickle
scikit-learn>=1.3.0,<1.5.0         # ← Limiter pour compatibilité joblib
joblib>=1.3.0,<1.4.0

# Utilities
pillow>=10.0.0